actix_web_lab

Module body

Experimental body types.

Analogous to the body module in Actix Web.

Structs

Sender

A channel-like sender for body chunks.

Writer

An AsyncWrite response body writer.

Functions

channel

Returns a sender half and a receiver half that can be used as a body type.

new_infallible_body_stream

Constructs a new BodyStream from an infallible byte chunk stream.

new_infallible_sized_stream

Constructs a new SizedStream from an infallible byte chunk stream.

writer

Returns an AsyncWrite response body writer and its associated body type.