actix_web_lab

Module extract

Experimental extractors.

Structs

BodyLimit

Extractor wrapper that limits size of payload used.

Bytes

Bytes extractor with const-generic payload size limit.

Host

Host information.

Json

JSON extractor with const-generic payload size limit.

LazyData

A lazy extractor for thread-local data.

LocalData

A thread-local equivalent to SharedData.

Path

Extract typed data from request path segments.

Query

Extract typed information from the request’s query.

ReconstructedPath

Reconstructed path using x-forwarded-prefix header.

RequestSignature

Wraps an extractor and calculates a request signature hash alongside.

SwapData

A wrapper around ArcSwap that can be used as an extractor.

UrlEncodedForm

URL-encoded form extractor with const-generic payload size limit.

Enums

RequestSignatureError

Errors that can occur when extracting and processing request signatures.

Constants

DEFAULT_BODY_LIMIT

Default body size limit of 2MiB.

DEFAULT_BYTES_LIMIT

Default bytes payload size limit of 4MiB.

DEFAULT_JSON_LIMIT

Default JSON payload size limit of 2MiB.

DEFAULT_URL_ENCODED_FORM_LIMIT

Default URL-encoded form payload size limit of 2MiB.

Traits

RequestSignatureScheme

Define a scheme for deriving and verifying some kind of signature from request parts.

Type Aliases

SharedData

An alias for actix_web::web::Data<T> with a more descriptive name.