Struct actix_web_lab::extract::Json

pub struct Json<T, const LIMIT: usize = DEFAULT_JSON_LIMIT>(pub T);

JSON extractor with const-generic payload size limit.

Json is used to extract typed data from JSON request payloads.

Extractor

To extract typed data from a request body, the inner type T must implement the serde::Deserialize trait.

Use the LIMIT const generic parameter to control the payload size limit. The default limit that is exported (DEFAULT_LIMIT) is 2MiB.

When const generics defaults are supported (due for Rust 1.59) it will be possible to omit the default limit.

use actix_web::{post, App};
use actix_web_lab::extract::{DEFAULT_JSON_LIMIT, Json};
use serde::Deserialize;

#[derive(Deserialize)]
struct Info {
    username: String,
}

/// Deserialize `Info` from request's body.
#[post("/")]
async fn index(info: Json<Info>) -> String {
    format!("Welcome {}!", info.username)
}

const LIMIT_32_MB: usize = 33_554_432;

/// Deserialize payload with a higher 32MiB limit.
#[post("/big-payload")]
async fn big_payload(info: Json<Info, LIMIT_32_MB>) -> String {
    format!("Welcome {}!", info.username)
}

Tuple Fields
0: T
Implementations

impl<T, const LIMIT: usize> Json<T, LIMIT>

pub fn into_inner(self) -> T

Unwrap into inner T value.