Struct actix_web_lab::header::Hsts

pub struct Hsts {
    pub include_subdomains: bool,
    pub preload: bool,
    /* private fields */
}

HTTP Strict Transport Security (HSTS) configuration.

Care should be taken when setting up HSTS for your site; misconfiguration can potentially leave parts of your site in an unusable state.

Default

The Default implementation uses a 5 minute max-age and does not include subdomains or preloading. This default is intentionally conservative to prevent accidental misconfiguration causing irrecoverable problems for users.

Once you have configured and tested the default HSTS config, recommended can be used as a secure default for production.

References

See the HSTS page on MDN for more information.

Fields
include_subdomains: bool

The includeSubdomains directive.

preload: bool

The preload directive.

Implementations

impl Hsts

pub fn new(duration: Duration) -> Self

Constructs a new HSTS configuration using the given duration.

Other values take their default.

pub fn recommended() -> Self

Constructs a secure, production-ready HSTS configuration.

Uses a max-age of 2 years and includes subdomains.

pub fn include_subdomains(self) -> Self

Send includeSubdomains directive with header.

pub fn preload(self) -> Self

Send preload directive with header.

See https://hstspreload.org/ for more information.