Module actix_web_lab::middleware

Experimental middleware.

Analogous to the middleware module in Actix Web.

Structs

ErrorHandlers

Middleware for registering custom status code based error handlers.

MiddlewareFn

Middleware transform for from_fn.

Next

Wraps the “next” service in the middleware chain.

RedirectHttps

A middleware to redirect traffic to HTTPS if connection is insecure.

Functions

from_fn

Wraps an async function to be used as a middleware.

redirect_to_www

A function middleware to redirect traffic to www. if not already there.