Crate actix_web_lab

In-progress extractors and middleware for Actix Web.

What Is This Crate?

This crate serves as a preview and test ground for upcoming features and ideas for Actix Web’s built in library of extractors, middleware and other utilities.

Any kind of feedback is welcome.

Complete Examples

See the examples folder for some complete examples of items in this crate.

Things To Know About This Crate

It will never reach v1.0.

Minimum Supported Rust Version (MSRV) is latest stable at the time of each release.

Breaking changes will likely happen on most 0.x version bumps.

Documentation might be limited for some items.

Items that graduate to Actix Web crate will be marked deprecated here for a reasonable amount of time so you can migrate.

Migrating will often be as easy as dropping the _lab suffix from imports when migrating.

Modules

body

Experimental body types.

extract

Experimental extractors.

guard

Experimental route guards.

header

Experimental typed headers.

middleware

Experimental middleware.

respond

Expiremental responders and response helpers.

sse

Semantic server-sent events (SSE) responder with a channel-like interface.

test

Expiremental testing utilities.

util

Utilities for working with Actix Web types.

web

Experimental services.

Macros

assert_response_matches

Quickly write tests that check various parts of a ServiceResponse.

test_request

Create a TestRequest using a DSL that looks kinda like on-the-wire HTTP/1.x requests.

Derive Macros

FromRequest

Derive a FromRequest implementation for an aggregate struct extractor.